requirements.txt
# Python 3.11 recommended
# Core for PDF diff
PyMuPDF>=1.24,<1.26
rapidfuzz>=3.9,<4
regex>=2024.4

# Core for DOCX diff
mammoth>=1.6,<2
beautifulsoup4>=4.12,<5
diff-match-patch>=20230430
python-docx>=1.1,<1.2

# --- Optional extras ---
# If you need OCR for scanned PDFs, uncomment the following:
# pytesseract>=0.3.10
# pdf2image>=1.17,<2
# pillow>=10.0,<11

# If you want to experiment with table-aware diffs in PDFs (needs system deps: ghostscript, opencv):
# camelot-py[cv]>=0.11,<0.12
